+++ docs/changelog.md
@@ -5,6 +5,8 @@ permalink: /changelog/
 ---
 
 # 147.0 (In Development)
+* **concept-engine** and **browser-engine-gecko**
+  * 🆕 New `verticalScrollPosition` and `verticalScrollDelta` APIs exposing the current scroll position and delta of the webpage [Bug 1990215](https://bugzilla.mozilla.org/show_bug.cgi?id=1990215)
 
 # 146.0